Разбил на модули, исправил вызов функции createAds() на вызов массива с объектами
