+++ js/create-ads.js
@@ -6,6 +6,8 @@ const CHECKOUT_TIMES = ['12:00', '13:00', '14:00'];
 const FEATURES_OF_PLACES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
 const PHOTO_OF_PLACES = ['https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg', 'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg', 'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'];
 
+const COUNT_SIMILAR_ADS = 10;
+
 const createAds = () => {
   const location = {
     lat: findRandomCoordinates(35.65000, 35.70000, 5),
@@ -19,7 +21,7 @@ const createAds = () => {
     location,
 
     offer: {
-      title: 'Квартира',
+      title: 'Аренда прекрасного места',
       adress: `${location.lat}, ${location.lng}`,
       price: findRandomIntegerNumber(500, 10000),
       type: TYPES_OF_HOUSES[findRandomIntegerNumber(0, TYPES_OF_HOUSES.length - 1)],
@@ -34,4 +36,6 @@ const createAds = () => {
   };
 };
 
-export const similarAds = (count) => new Array(count).fill(null).map(() => createAds());
+const similarAds = () => new Array(COUNT_SIMILAR_ADS).fill(null).map(() => createAds());
+
+export { similarAds };
